+++ src/otx/algorithms/visual_prompting/tasks/openvino.py
@@ -65,7 +65,7 @@ from otx.api.entities.task_environment import TaskEnvironment
 from otx.api.serialization.label_mapper import LabelSchemaMapper, label_schema_to_bytes
 from otx.api.usecases.evaluation.metrics_helper import MetricsHelper
 from otx.api.usecases.exportable_code import demo
-from otx.api.usecases.exportable_code.inference import BaseInferencer
+from otx.api.usecases.exportable_code.inference.inference import IInferencer
 from otx.api.usecases.exportable_code.prediction_to_annotation_converter import (
     VisualPromptingToAnnotationConverter,
 )
@@ -80,7 +80,7 @@ from otx.api.usecases.tasks.interfaces.optimization_interface import (
 logger = get_logger()
 
 
-class OpenVINOVisualPromptingInferencer(BaseInferencer):
+class OpenVINOVisualPromptingInferencer(IInferencer):
     """Inferencer implementation for Visual Prompting using OpenVINO backend.
 
     This inferencer has two models, image encoder and decoder.